+++ stacks-common/src/types/mod.rs
@@ -790,6 +790,29 @@ impl StacksEpochId {
         }
     }
 
+    /// Before Epoch 3.3, the cost for arguments to functions was based on the
+    /// parameter type, not the actual size of the argument passed in. This
+    /// resulted in over-charging for arguments smaller than the maximum size
+    /// permitted for the parameter.
+    pub fn uses_arg_size_for_cost(&self) -> bool {
+        match self {
+            StacksEpochId::Epoch10
+            | StacksEpochId::Epoch20
+            | StacksEpochId::Epoch2_05
+            | StacksEpochId::Epoch21
+            | StacksEpochId::Epoch22
+            | StacksEpochId::Epoch23
+            | StacksEpochId::Epoch24
+            | StacksEpochId::Epoch25
+            | StacksEpochId::Epoch30
+            | StacksEpochId::Epoch31
+            | StacksEpochId::Epoch32 => false,
+            StacksEpochId::Epoch33 => true,
+        }
+    }
+
+    /// In Epoch 3.3, limits are introduced on the number of parameters
+    /// in function definitions and the number of methods in trait definitions.
     pub fn limits_parameter_and_method_count(&self) -> bool {
         match self {
             StacksEpochId::Epoch10